refactor: viewmodel refactor finished.

+++ app/src/main/java/com/example/shopping_list/viewmodel/CartViewModel.kt
@@ -5,6 +5,7 @@ import com.example.shopping_list.model.Product
 import com.example.shopping_list.repository.CartRepository
 import com.example.shopping_list.repository.FavoriteRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.StateFlow
 import javax.inject.Inject
 
 
@@ -15,12 +16,14 @@ class CartViewModel @Inject constructor(
 ): ViewModel(){
     val cartItems=cartRepository.cartItems
 
-    fun addToFavorites(product: Product){
-        favoriteRepository.addProductToFavorites(product)
-    }
+    val favoriteItems: StateFlow<List<Product>> = favoriteRepository.favoriteItems
 
-    fun removeFromFavorites(product: Product){
-        favoriteRepository.removeProductFromFavorites(product)
+    fun toggleFavorite(product: Product) {
+        if (favoriteRepository.isFavorite(product)) {
+            favoriteRepository.removeProductFromFavorites(product)
+        } else {
+            favoriteRepository.addProductToFavorites(product)
+        }
     }
 
     fun removeFromCart(product: Product){
